Other/agro_ai/requirements.txt
numpy
pandas
scikit-learn
joblib
tensorflow
rasterio
requests
matplotlib
Pillow
scipy
folium
apscheduler
